Moved AboutController into the controllers folder
